apps/ssh-manager/requirements.txt
PyYAML==5.3.1
requests==2.23.0